+++ src/bible/routeUtils.nim
@@ -1,14 +1,6 @@
 from std/logging import nil
-
 from std/json import parseJson, `{}=`, `%`, newJObject
-
 from std/strtabs import keys
-
-import bible/config
-export config
-import bible/utils
-export utils
-
 
 const
   autoFormParsing {.boolDefine.} = true
@@ -18,8 +10,14 @@ when autoXmlParsing:
   from std/xmlparser import parseXml
   import std/xmltree
 
+import bible/config
+export config
+import bible/utils
+export utils
+
 import pkg/prologue
 export json
+
 
 proc hasKey*(session: var Session; key: string): bool =
   ## Check if session have specific session
@@ -38,14 +36,14 @@ proc setContentJsonHeader*(ctx) =
   ## Set the response content-type to json
   ctx.response.setHeader "content-type", "application/json"
 
-template withParams*(ctx; mergeGet = false; mergePath = false; bodyCode: untyped) =
+template withParams*(ctx; get = false; path = false; bodyCode: untyped) =
   ## Run `body` if request has a JSON body
   ##
   ## If no JSON sent or/and the `content-type` is not
   ## JSON, it will response with `Http400` and close
   ## connection
   ##
-  ## Use `mergeGet` to merge the get parameters into `node`
+  ## Use `get` to merge the get parameters into `node`
   ## The GET parameters override the POST
   let reqMethod = ctx.request.reqMethod
   var
@@ -69,24 +67,16 @@ template withParams*(ctx; mergeGet = false; mergePath = false; bodyCode: untyped
       when autoXmlParsing:
         {.fatal: "XML parsing not implemented".}
     else: discard
-  if mergeGet or reqMethod == HttpGet:
+  else:
+    error = false
+  if get or reqMethod == HttpGet:
     for key, val in ctx.request.queryParams:
       node{key} = %val
-  if mergePath:
+  if path:
     for key, val in ctx.request.pathParams:
       node{key} = %val
   logging.debug "Auto parsed params: " & $node
   bodyCode
-
-template withParams*(ctx; mergeGet = false; bodyCode: untyped) =
-  ## Alias for `ctx.withParams(mergeGet, false): bodyCode`
-  ctx.withParams(mergeGet, false): bodyCode
-template withParams*(ctx; mergePath = false; bodyCode: untyped) =
-  ## Alias for `ctx.withParams(mergeGet, false): bodyCode`
-  ctx.withParams(false, mergePath): bodyCode
-template withParams*(ctx; bodyCode: untyped) =
-  ## Alias for `ctx.withParams(false, false): bodyCode`
-  ctx.withParams(false, false): bodyCode
 
 type
   ResponseJson* = object
@@ -104,10 +94,12 @@ func initResponseJson(
 ): ResponseJson =
   ResponseJson(kind: kind, text: text, error: error)
 
-  
 template respJson*(data: untyped; code: HttpCode) =
   ## Send a JSON to client
-  resp($(%*data), code)
+  when logicInApi:
+    resp($(%*data), code)
+  else:
+    resp(data.text, code)
 
 template respErr*(msg: string; code = Http400) =
   ## Send a error message in a JSON to client
@@ -217,3 +209,15 @@ proc getUsing*(
       let val = node{field}.getStr
       result = table.get(val, [inDb])
       break
+
+import bible/views/error/docNotExists
+
+template withDoc*(ctx; doc: string; body: untyped): untyped =
+  ## Check if the document exists
+  if doc in getAllDocsName():
+    body
+  else:
+    ctx.render docNotExists doc
+
+  
+