+++ src/bible/routes/versicles.nim
@@ -7,22 +7,17 @@ import bible/db/models/document
 import bible/routeUtils
 
 import bible/views
-import bible/views/[
-  docNotExists,
-  versicles
-]
+import bible/views/versicles
 
 proc r_versicles*(ctx: Context) {.async.} =
   ## Versicles
   ctx.forceHttpMethod HttpGet
-  ctx.withParams(mergePath = true):
+  ctx.withParams(get = false, path = true):
     node.ifContains(all = ["doc", "book", "verse"]):
       let
         doc = node{"doc"}.getStr
         book = node{"book"}.getStr
         verse = node{"verse"}.getInt
 
-      if doc in getAllDocsName():
+      ctx.withDoc doc:
         ctx.render versicles(doc, book, verse, @["verse1", "verse2", "verse3"])
-      else:
-        ctx.render docNotExists doc