+++ src/bible/views/books.nim
@@ -8,16 +8,16 @@ import pkg/karax/[
 from bible/views import View
 from bible/utils import url
 
-proc verses*(doc, book: string; verses: openArray[int]): View =
+proc books*(doc: string; books: openArray[string]): View =
   result.code = Http200
-  result.name = fmt"{doc} - {book} Verses"
+  result.name = fmt"{doc} books"
   result.vnode = buildHtml(tdiv):
     h1:
       text result.name
-    tdiv(class = "verses"):
-      for verse in verses:
-        tdiv(class = "verse"):
-          a(href = url fmt"/{doc}/{book}/{verse}"): text $verse
+    tdiv(class = "books"):
+      for book in books:
+        tdiv(class = "book"):
+          a(href = url fmt"/{doc}/{book}"): text book
 
     
     # script(src = "script/home.js") # getJs "home"